Deprecate problems key in config.json
Until now the config.json has used the "problems" key to
define which exercises are in the track, and in which order.

The problems key was a simple array of strings, each of which
represents a problem slug.

E.g.:

    {
      "problems": [
        "hello-world",
        "leap",
        "acronym"
      ]
    }

We are now deprecating this key and reading from the new "exercises" key,
which contains an array of objects. Each object contains the following
three keys:

- `slug`
- `difficulty`
- `topics`

E.g.

    {
      "exercises": [
        {
          "slug": "hello-world" ,
          "difficulty": 1,
          "topics": [
              "control-flow (if-statements)",
              "optional values",
              "text formatting"
          ]
        },
        {
          "difficulty": 3,
          "slug": "anagram",
          "topics": [
              "strings",
              "filtering"
          ]
        },
        {
          "difficulty": 10,
          "slug": "forth",
          "topics": [
              "parsing",
              "transforming",
              "stacks"
          ]
        }
      ]
    }

If the exercises key is present, the problems key will be ignored. If not
we will fall back to reading from the problems key.