Remove support for deprecated problem metadata
Previously we've had all the exercise metadata files in the root folder.

```
.
├── acronym.json
├── acronym.md
└── acronym.yml
```

We decided to clean up the x-common repository, changing the format to
the following:

```
.
└── exercises
  └── acronym
    ├── canonical-data.json
    ├── description.md
    └── metadata.yml
```

We added support to read from both formats so that we could swap
seamlessly in production. Now that the x-common repository has been
updated we don't need support for the deprecated structure anymore.