Implement more granular state checks
For tracks that are not active we distinguish between tracks that are upcoming (they've got
at least one exercise implemented) and planned (which someone asked for, but no work has
been done on it yet).

We've been doing extra filtering and checking on the client side to handle this. With this
change the client-code can be simplified a bit.