Fix confusion around domain concepts.
The 'problems' method on the track was actually returning slugs.

Yeah, really confusing.

* _slug_ - the URL-friendly identifier for an exercise.
* _problem_ - the generic, language-independent metadata for an exercise.
* _implementation_ - the language-specific implementation of an exercise.

The domain concepts are decent, provided we can get all the variables and
methods named consistently.

I think the origin of the confusion was that the configuration file
refers to a number of different sets of slugs:

- active exercises in the track
- foregone exercises (which will never be implemented)
- deprecated exercises (implemented, but no longer active)

So we were referring to the full set of slugs as "slugs", and needed
something to call the thing that represents "active slugs", and since
these used to be implemented in a key called "problems" in the config
file, it was probably a very short road to having a "problems" method
that simply returned the array directly from the config.

Then we deprecated the problems array in the config, to have a richer
data set for the active exercises for the track, under the "exercises"
key... and we couldn't just call the active slugs "exercises", because
that has more data.

To clear this up I've added more context to each of the lists of
slugs (active slugs, foregone slugs, and deprecated slugs), still
referring to the combined set as just plain "slugs". The active
slugs method uses some intentionally horribly named (but private)
methods, which at least makes it completely explicit what is going on.

Finally, I added a convenience method for problems that actually returns
problems, since we were previously accessing problems via the
implementations.

wip