Move `deprecated?` responsiblity into `Problem`.
`Problem` already knows about 3 of the 4 possible problem related files
that can live in the exercise directory.

`metadata.yml`,`description.md`,`canonical-data.json`

This patch adds `.deprecated` so that they are now all covered.
And a `deprecated?` method to check if a problem is deprecated.

`Problems.deprecated?` was a private method so removing it should be safe.