Reapply "implementation: Test for symlinked files"
"implementation: Test for symlinked files" (of PR #12) was reverted
because PR #17 caused the fixtures to be in an inconsistent state.

Now that the previous commit has restored consistency, we can reapply
the commit of #12.

The rationale for reapplying is the same as that of in #12: We would
like to ensure that support for symlinks remains (as it currently
exists).