implementation: Test for symlinked files
We have a use case in https://github.com/exercism/xhaskell/issues/442
where we would like to symlink files.

We have one common file that will get delivered with each exercise, and
we would only like to have to maintain one instance of it in the repo,
while symlinking it into all exercise directories.

This is a separate case than https://github.com/exercism/x-api/issues/29
because that talks about files that are not associated with any
particular exercise - their destination is a directory outside of any
exercise directory, whereas our destination is inside each exercise
directory.

For this case to work for us, we need trackler to correctly follow
symlinks present in an implementation. It currently does. In adding this
test, we request that this behaviour in the face of a symlink not
change in future trackler versions, for we have started to depend on it
as of https://github.com/exercism/xhaskell/pull/443