Add support for a TRACK_HINTS.md file
The TRACK_HINTS.md file is included into the README of every exercise
in the track. If TRACK_HINTS.md is missing, it will use SETUP.md. If
neither exist, then the README will be generated without any track-level
hints.

This is in support of deprecating the SETUP.md file, which is terribly
confusingly named. See exercism/exercism.io#3240