Make documentation image paths configurable
In order to avoid hard-coding complete API paths for images, we need to add
a way to let the caller define where the documentation images can be found.

This also lets us see the images in the markdown on GitHub by defaulting to the
./docs/img path location.

It's kind of magical, but it makes a lot more sense than what we've been doing
and it makes it easier to contribute.

See exercism/exercism.io#3312