fixtures: Move exercises into exercises directory
PR #17 (replacing PR #14) created `animal/exercises` and
`jewels/exercises`.

There is an assumption that either a track has the `exercises` directory
and **all** its exercises are in that directory, or else it has no
`exercises` directory.

These two tracks were made to violate this assumptions by the creation
of the `exercises` directory; they have exercises outside that
directory.

This caused a conflict on merge of #17 and #12. There was no merge
conflict in the `git` sense, but the tests for #17 and #12 both
independently pass, while the tests will fail on the merged result.

We can make everything consistent again by moving the exercises into the
`exercises` directory for these tracks.