Refactor to expose the two root concepts
No functional or API changes here. Just a refactoring of the internals
of Problem.

With just `@root` there was a some extra effort necessary to switch the
various file/paths into either URLs or filesystem locations. This can be
simplified by clearly splitting the concepts.

Within a problem there is now a file_root and a repo_root. The former is
used when looking at a filesystem, the latter for creating URLs.

It's possible that these two responsibilities should live elsewhere. But
that's for later.

The various files contained within a problem have had their `_path`
methods replaced with a `_file_name` method that can be used to build
either filesystem locations or URLs.

The `file_path` and `repo_url` return what you'd expect.