Add test track with custom problem
The snowflake track implements a custom problem, snowflake-only.

The test of Tracks had to change to adjust to the new fixture. This test
might be a bit brittle, but that's something to address later.

Adding e new test of Track to make sure that its `problems` and
`implementations` contain a the `snowflake-only` problem.

These tests pass with no code changes, so there are no functional
changes in this commit.

However, the problem returned by these methods is wrong. If you ask the
location of its files it will point you to `common`, and that is wrong.

However, the Track test probably isn't the place to test that. At least
not right now. I think that will change.