Shameless green of a track-specific problem
Yeah, don't tell me that this code is ugly. I'm aware.

This commit establishes tests for how a problem should behave if it has
track-specific files.

Note that this has _interesting_ side effects. If a track were to define
their own custom metadata.yml file for a problem that exists in common,
then that custom metadata.yml file would be used over the commonly
defined one.

This may not be what we want. We'll have to figure that out.

Once we're comfortable with how to handle these situations this code
will all be refactored. Which is why I'm not worried about the ugliness.