Extract Metadata creation
Probably a bigger refactor than was strictly necessary. But this
(mostly) extracts the metadata concept from Problem and puts it into a
Metadata object

The `Metadata.for` method builds an appropriate metadata object. The
options are:

- Metadata using a file in the track
- Metadata using a file in common
- A NullMetadata object if no metadata.yml file exists for the problem

The `metadata.exists?` thing is lifted from Track, which already uses
that predicate.

Problem previously defaulted to `nil` for the track param, but that
doesn't work once I wanted to pass the track on to Metadata, so I've
also introduced NullTrack.

A nice thing about having a guaranteed track object is that I can call
`track.exists?` before building the track-specific ivars.