Separate description from its file
I wanted to remove the triple-if from Description#for, this led me to
removing the file behaviors from Description entirely. Description now
always returns a Description (before it returned either a Description or
a NullDescription)

The part of Description that varies, its content and where it comes
from, is now down in the DescriptionFile logic.

I've put all 4 DescriptionFile variants together in a file. I think if
someone wants to look at one they'll probably want to look at the
others. The classes work together, not independently.

Now that Description#for no longer returns a NullDescription, the
implementation of the `exists?` method needed to change. It now checks
for the presence of content. If there is no description file, the
content will be empty-string