Extract metadata_file from metadata
I thought this would be identical to Description/Description file, but
there are a few differences. MetadataFile has `to_h`, while Description
has `to_s`. Because the metadata file is structured data, while the
description is just text.

Other than that they seem to be pretty much the same and there's
probably a concept that can be extracted out of this.