Move github url to File objects
The last bit of description/metadata knowlegde still living in problem
was the Github URL code. This commit moves that knowledge into
Metadata and Description objects.

As a bonus, I no longer have to do the weird `track_repo_root` ivar and
the associated conditionals in Problem. The remaining clients of
`repo_root` method only work with files that exist in x-common, so I no
longer have to check with github url root to use.