Pass track to Problem when creating implementations
Passes the test to show a track can return track-specific problems.

Now that we can get track-specific problems by providing a track to
`Problem.new`, we want to make sure to create track-specific problems
when building our Implementations collection for a track.

Doing this requires passing the track into Implementations so that it
can be passed on to Problem.

This changed the arity of Implementations.new, so that test needed an
update. But there are no clients of `Implementations.new` outside of
this gem, so this is not a breaking change.