Test final pieces of Trackler behavior
As mentioned in #22

```
Trackler.problems

Current behavior: Returns a Problems collection that contains all
problems in the common directory

Proposed behavior: Unchanged

Would it contain track-specific problems? No.
```

and

```
Trackler.implementations

Current behavior: Returns a Hash. The key is a problem slug, the values
is an array of all tracks that implement that problem.

Proposed behavior: Same structure, but if a track implements a
track-only problem, it would be contained in this Hash. So, in the
example we would now have

{..."rust-only" => [<rust implementation object>]...}

Trackler.implementations['rust-only'] would return a an array with a
single implementation object.a
```

and

```
Trackler.todos

Current behavior: Returns a Hash. The key is a track, the value is an
array of common problems that the track does not implement.

Proposed behavior: Unchanged

Would it contain track-specific problems? No.

Trackler.todos['rust-only'] would return nil.
```

These tests exercise these behaviors