Add tests for DescriptionFile, MetadataFile
The behavior of these is tested through Problem, but, hey, why not test
these files directly.

I only wanted to test the behavior of `for` and the state of the objects
it creates, so I've made a bunch of methods private, leaving only the
necessary API public. This does not change the behavior of these
classes, only narrows how they can be used down to the ways they
_should_ be used.