Allow tracks to specify their own ignore_pattern
If a file matches the ignore pattern it will be removed from the
implementation's file bundle.

If a track does not provide an `ignore_pattern` in its `config.json`
file the default of `/example/` will be used.

This is necessary since the Go track is implementing example tests,
which have example in the file name, but are not part of the reference
solution.