Suppress blurb if it starts the description
https://github.com/exercism/x-common/pull/775 moves the first line of
the description of an implementation from the blurb to the description
so that the description will be self contained.

If Trackler prepends the blurb to the description when it is already
there, the blurb will appear to be doubled up in the readme.

This patch checks, and does not add the blurb to the readme if it
matches the start of the description text.