Remove ability to mutate Implementation#@files
Being able to mutate the files leads to bugs where the files you think
you have are not the files you actually have.

Remove the temptation to ever do this by removing support for it.

The removed methods `#files=`, `merge_files` are both now unused since
(x-api) was rewritten to avoid them.