Update the update script (again)
This time try to make sure that each of the submodules are actually ON master when
updating. They usually are, but you know how it goes. Something goes wrong downstream,
we manually go in and check out some other SHA that we know we can trust, and then we
forget to go back laster. We being me. Anyway. Here we go (again).