Update fixture to weight doc format more heavily
The 'fruit' track fixture contained two files in the user-facing docs directory.
Each of these files had different extensions, causing the outcome of 'most_popular_format' to be
ambiguous.

In the real world, if the result is ambiguous, we're OK with it (we're moving away from this implementation,
for one, and for the other, we can't really know which is the right one if there are equal numbers of two
different ones).

For the sake of the tests, it's more useful to have a file format that 'wins'.